im_translate change and use utest_tolerance for image shift checks

+++ src/aspire/image/image.py
@@ -341,7 +341,8 @@ class Image:
 
         # Note original stack shape and flatten stack
         stack_shape = self.stack_shape
-        im = self.stack_reshape(-1)._data
+        # upcast to np.float64 to preserve numerical stability in FFT ops
+        im = self.stack_reshape(-1)._data.astype(np.float64)
 
         if shifts.ndim == 1:
             shifts = shifts[np.newaxis, :]
@@ -357,9 +358,7 @@ class Image:
 
         L = self.resolution
         im_f = xp.asnumpy(fft.fft2(xp.asarray(im)))
-        grid_shifted = fft.ifftshift(
-            xp.asarray(np.ceil(np.arange(-L / 2, L / 2, dtype=self.dtype)))
-        )
+        grid_shifted = fft.ifftshift(xp.asarray(np.ceil(np.arange(-L / 2, L / 2))))
         grid_1d = xp.asnumpy(grid_shifted) * 2 * np.pi / L
         om_x, om_y = np.meshgrid(grid_1d, grid_1d, indexing="ij")
 
@@ -373,7 +372,8 @@ class Image:
         mult_f = np.exp(-1j * phase_shifts)
         im_translated_f = im_f * mult_f
         im_translated = xp.asnumpy(fft.ifft2(xp.asarray(im_translated_f)))
-        im_translated = np.real(im_translated)
+        # downcast back to Image's internal dtype
+        im_translated = np.real(im_translated).astype(self.dtype)
 
         # Reshape to stack shape
         return Image(im_translated).stack_reshape(stack_shape)
